+++ internal/cmd/drc/devices_remove.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/spf13/cobra"
 
+	"github.com/inhandnet/devicemanager-cli/internal/api"
 	"github.com/inhandnet/devicemanager-cli/internal/factory"
 	"github.com/inhandnet/devicemanager-cli/internal/iostreams"
 	"github.com/inhandnet/devicemanager-cli/internal/ui"
@@ -32,7 +33,9 @@ func NewCmdDevicesRemove(f *factory.Factory) *cobra.Command {
 
 			output, _ := cmd.Flags().GetString("output")
 
-			resp, err := client.Delete(fmt.Sprintf("/api/job/%s/devices/%s", templateID, deviceID))
+			resp, err := client.Do("DELETE", fmt.Sprintf("/api/job/%s/devices/%s", templateID, deviceID), &api.RequestOptions{
+				Query: oidQuery(f),
+			})
 			if err != nil {
 				return err
 			}